[#32785250] Сопоставил классы ProfileInfoEntity и IntentEntity с таблицами в базе данных. Пометил поля в классах nullable в соответствии с полями в таблице. Not nullable поля инициализирую через конструктор. Там где необходимо провожу каст из nullable в not nullable.

+++ LeokaEstetica.Platform.Finder/Services/Resume/ResumePaginationService.cs
@@ -81,7 +81,7 @@ public class ResumePaginationService : BaseIndexRamDirectory, IResumePaginationS
                     var userId = res.UserId;
                     
                     // Проверяем заполнение анкеты.
-                    var isEmptyProfile = await _accessUserService.IsProfileEmptyAsync(userId);
+                    var isEmptyProfile = await _accessUserService.IsProfileEmptyAsync((long)userId);
 
                     // Удаляем анкеты из выборки, которые не заполнены.
                     if (isEmptyProfile)
